% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gb_agg_ratio}
\alias{gb_agg_ratio}
\title{Group by ratio}
\usage{
gb_agg_ratio(df, groupvar, numerator_var, denominator_var, aggfun, valuename,
  varname)
}
\arguments{
\item{df}{the data frame containing the variables}

\item{groupvar}{the variable to group by for ratio compare}

\item{numerator_var}{the vector of variables as the numerators}

\item{aggfun}{the aggregation function, e.g. sum, mean}
}
\description{
This function first group by some variable A, compute some aggregation of the
numerator variables, take ratio of each numerator variables with the denominator_var
}
\examples{
gb_agg_ratio(df, groupvar = "Fiscal_Year", numerator_var = c("Electricity_(Cost)", "Gas_(kBtu)", "Oil_(kBtu)", "Steam__(kBtu)", "Chilled_Water_(kBtu)"), denominator_var = "Gross_Sq.Ft")
}
\keyword{by}
\keyword{group}
\keyword{plot}
\keyword{ratio}
