% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_filter_set}
\alias{get_filter_set}
\title{Get the subject building set}
\usage{
get_filter_set(category, type, year, region)
}
\arguments{
\item{category}{optional, a subset of A, B, C, D, E, I to include}

\item{type}{optional, a string (e.g. "Office"), or a string vector (e.g. c("Office", "Courthouse")) of building type}

\item{year}{optional, the year to plot}

\item{region}{optional, the region to plot}
}
\description{
This function gets the subject building set
}
\examples{
get_filter_set(category=c("A", "I"), year=2017, region="9")
}
\keyword{count}
\keyword{query}
