% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{median_summary}
\alias{median_summary}
\title{Summary table own median vs cbecs median}
\usage{
median_summary()
}
\description{
This function generates a data frame with building type, cbecs median, region 1 median, ..., region 11 median
}
\examples{
dollar_saving(category=c("I", "A"), year=2017, region="9")
}
\keyword{median}
\keyword{summary}
