% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dollar_saving}
\alias{dollar_saving}
\title{Potential dollar saving based on median}
\usage{
dollar_saving(category, type, year, region, method = "own")
}
\arguments{
\item{category}{optional, a subset of A, B, C, D, E, I to include}

\item{type}{optional, a string (e.g. "Office"), or a string vector (e.g. c("Office", "Courthouse")) of building type}

\item{region}{optional, the region to plot}

\item{years}{optional, the years to plot}

\item{reference}{optional, cbecs, own, or hybrid}
}
\description{
This function plots the potential dollar savings based on some median eui
}
\examples{
dollar_saving(category=c("I", "A"), year=2017, region="9")
}
\keyword{dollar}
\keyword{median}
\keyword{saving}
