% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{stackbar}
\alias{stackbar}
\title{Stacked bar plot}
\usage{
stackbar(df, xcol, fillcol, ycol, orderByHeight, ylabel, xlabel, tit, legendloc,
  legendOrient, pal, pal_values, labelFormat, width, verbose)
}
\arguments{
\item{df}{required, a dataframe to plot}

\item{xcol}{the horizontal variable}

\item{fillcol}{the fill column}

\item{ycol}{the y column}

\item{ylabel}{required, label for y axis}

\item{xlabel}{required, label for x axis}

\item{tit}{plot title}

\item{legendloc}{legend location: top, bottom, c(0.8, 0.2)}

\item{legendOrient}{legend orientation: "v" or "h"}

\item{pal}{color palette}

\item{pal_values}{colors in the palette to use}

\item{labelFormat}{optional, string format function of label}

\item{width}{optional, the width of the bars}

\item{aggfun}{count or sum}

\item{width}{verbose, if TRUE print data to the output}
}
\description{
This function produces a stacked bar plot
}
\examples{
stackbar(df=df, xcol="Fiscal_Year", fillcol="Cat", ylabel="Building Count", tit="EUAS Building Count By Category")
}
\keyword{count}
\keyword{query}
